move /fixes/* to /scripts/fixes
