+++ package.json
@@ -1,22 +1,21 @@
 {
-    "name": "jsquarto",
-    "version": "1.0.0",
+    "name": "@oscsa/jsquarto",
+    "version": "1.0.2",
     "description": "Generate JS package API reference documentation using Markdown and Quarto. JSquarto is designed as an alternative to JSDoc",
     "main": "index.js",
     "scripts": {
         "doc:generate": "ts-node-dev --transpile-only --project tsconfig.json ./src/index.ts",
         "doc:preview": "quarto preview ./docs",
-        "fix:all": "ts-node-dev --transpile-only --project tsconfig.json ./src/fixes/index.ts",
-        "fix:crowdin_files": "ts-node-dev --transpile-only --project tsconfig.json ./src/fixes/fix_crowdin_translation.ts",
-        "fix:localized_files": "ts-node-dev --transpile-only --project tsconfig.json ./src/fixes/fix_localized_index_file.ts",
-        "fix:duplicate_languages_ref": "ts-node-dev --transpile-only --project tsconfig.json ./src/fixes/fix_duplicate_language_refs.ts",
-        "fix:fix_wrong_language_refs": "ts-node-dev --transpile-only --project tsconfig.json ./src/fixes/fix_wrong_language_refs.ts",
+        "fix:all": "ts-node-dev --transpile-only --project tsconfig.json ./src/scripts/fixes/index.ts",
+        "fix:crowdin_files": "ts-node-dev --transpile-only --project tsconfig.json ./src/scripts/fixes/fix_crowdin_translation.ts",
+        "fix:localized_files": "ts-node-dev --transpile-only --project tsconfig.json ./src/scripts/fixes/fix_localized_index_file.ts",
+        "fix:duplicate_languages_ref": "ts-node-dev --transpile-only --project tsconfig.json ./src/scripts/fixes/fix_duplicate_language_refs.ts",
+        "fix:fix_wrong_languages_ref": "ts-node-dev --transpile-only --project tsconfig.json ./src/scripts/fixes/fix_wrong_language_refs.ts",
         "doc:clean": "rm -rf ./docs",
         "doc:serve": "serve ./docs/_book",
         "setup-quarto": "git clone https://github.com/quarto-dev/quarto-cli && cd quarto-cli && ./configure.sh",
         "update-quarto": "cd quarto-cli && git pull",
-        "build": "npm run doc:generate && npm run doc:preview",
-        "tsc": "tsc -b && ts-node-dev ./src/add-shebang.ts"
+        "tsc": "tsc -b && ts-node-dev ./src/add_shebang.ts"
     },
     "bin": {
         "jsq": "./bin/src/jsq.js"
@@ -27,7 +26,6 @@
     "dependencies": {
         "acorn": "^8.11.3",
         "cheerio": "^1.0.0-rc.12",
-        "cross-env": "^7.0.3",
         "fs": "^0.0.1-security",
         "serve": "^14.2.3",
         "ts-node-dev": "^2.0.0",
