+++ src/jsq.ts
@@ -20,7 +20,6 @@ const commandsWithNoArgs = [
     "update-quarto",
 ];
 
-// const additionalArgs = args.slice(1).join(" ") + " workingDir=" + process.cwd();
 // Determine the jsq tool directory
 const projectDir = path.resolve(__dirname, "..");
 const args = process.argv.slice(2);
@@ -34,7 +33,9 @@ if (!allowedCommands.includes(script)) {
 
 const command = commandsWithNoArgs.includes(script)
     ? script
-    : `${script} -- ${additionalArgs}`;
+    : `${script} -- ${additionalArgs}  workingDirectory=${process.cwd()}`;
 
 process.chdir(projectDir);
-execSync(`npm run ${command}`, { stdio: "inherit" });
+execSync(`npm run ${command}`, {
+    stdio: "inherit",
+});
