+++ src/scripts/add_shebang.ts
